chore: moving modules logic to a specific file
